pr-analyser-pat/pyproject.toml
[tool.poetry]
name = "pr-analyser"
version = "0.1.0"
description = "GitHub PR and Commit Analysis Tool"
authors = ["Jeremy Chia <jeremy@example.com>"]
readme = "README.md"
packages = [{include = "pr_analyser"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
keyring = "^25.6.0"
pandas = "^2.1.0"
matplotlib = "^3.8.0"
plotly = "^5.17.0"
pyyaml = "^6.0.1"
rich = "^13.7.0"
pygithub = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
pr-analyser = "pr_analyser.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"